letchange1

+++ product/models.py
@@ -16,7 +16,7 @@ class Product(models.Model):
     description = models.TextField(blank=True, null=True)
     slug = models.SlugField(max_length=255)
     
-    price = models.DecimalField(max_digits=6, decimal_places=2)
+    price = models.PositiveIntegerField
     date_added = models.DateTimeField(auto_now_add=True)
     image = models.ImageField(upload_to='uploads/', blank=True, null=True)
     thumbnail = models.ImageField(upload_to='uploads/', blank=True, null=True)
